feat: split flash and non flash grammar tests
